// idle-buds/src/enums/game.enums.ts
export enum GameMode {
    ADVENTURE = 'ADVENTURE',
    HARDCORE = 'HARDCORE',
    NUZLOCKE = 'NUZLOCKE'
  }
  
  export enum ProgressionTier {
    TIER_1 = 'TIER_1',
    TIER_2 = 'TIER_2',
    TIER_3 = 'TIER_3',
    TIER_4 = 'TIER_4',
    TIER_5 = 'TIER_5',
    TIER_6 = 'TIER_6',
    TIER_7 = 'TIER_7',
    TIER_8 = 'TIER_8',
    ASCENDED = 'ASCENDED'
  }